// src/models/Renpho.ts
interface RenphoDataModel {
  time_of_measurement: Date;
  weight: number;
  bmi: number;
  body_fat: number;
  fat_free_body_weight: number;
  subcutaneous_fat: number;
  visceral_fat: number;
  body_water: number;
  skeletal_muscle: number;
  muscle_mass: number;
  bone_mass: number;
  protein: number;
  bmr: number;
  metabolic_age: number;
}

export default RenphoDataModel;